Moving I/O (aka communication) layer outside powHSM code
- The powHSM logic now handles only one command at a time
- Each implementation (i.e., Signer and TCPSigner) implement their own I/O and looping
- The communication module only facilitates access to the APDU buffer
