+++ firmware/src/hal/src/ledger/communication.c
@@ -1,49 +1,1 @@
-/**
- * The MIT License (MIT)
- *
- * Copyright (c) 2021 RSK Labs Ltd
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to
- * deal in the Software without restriction, including without limitation the
- * rights to use, copy, modify, merge, publish, distribute, sublicense, and/or
- * sell copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
- * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
- * IN THE SOFTWARE.
- */
-
-#include "os.h"
-#include "hal/communication.h"
-
-static unsigned char* msg_buffer;
-static size_t msg_buffer_size;
-
-// HAL implementation
-bool communication_init(unsigned char* _msg_buffer, size_t _msg_buffer_size) {
-    // Setup the exchange buffer
-    msg_buffer = _msg_buffer;
-    msg_buffer_size = _msg_buffer_size;
-    return true;
-}
-
-unsigned char* communication_get_msg_buffer() {
-    return msg_buffer;
-}
-
-size_t communication_get_msg_buffer_size() {
-    return msg_buffer_size;
-}
-
-unsigned short communication_io_exchange(unsigned short tx) {
-    return io_exchange(CHANNEL_APDU, tx);
-}
+../x86/communication.c
